Add basic hook
